Move BulkTagger to its own module
